feat: highlight the setting page
